+++ src/revlanguage/functions/math/Func_discretizeBetaQuadrature.cpp
@@ -31,7 +31,7 @@ RevBayesCore::TypedFunction< RevBayesCore::MatrixReal >* Func_discretizeBetaQuad
     
     RevBayesCore::TypedDagNode<double>* alpha = static_cast<const Real &>( this->args[0].getVariable()->getRevObject() ).getDagNode();
     RevBayesCore::TypedDagNode<double>* beta = static_cast<const Real &>( this->args[1].getVariable()->getRevObject() ).getDagNode();
-    RevBayesCore::TypedDagNode<int>* numCats = static_cast<const Integer &>( this->args[2].getVariable()->getRevObject() ).getDagNode();
+    RevBayesCore::TypedDagNode<long>* numCats = static_cast<const Integer &>( this->args[2].getVariable()->getRevObject() ).getDagNode();
 
     RevBayesCore::DiscretizeBetaQuadratureFunction *func = new RevBayesCore::DiscretizeBetaQuadratureFunction( alpha, beta, numCats);
