+++ src/revlanguage/functions/math/Func_discretizeLognormalQuadrature.cpp
@@ -31,7 +31,7 @@ RevBayesCore::TypedFunction< RevBayesCore::MatrixReal >* Func_discretizeLognorma
     
     RevBayesCore::TypedDagNode<double>* m   = static_cast<const Real &>( this->args[0].getVariable()->getRevObject() ).getDagNode();
     RevBayesCore::TypedDagNode<double>* s   = static_cast<const RealPos &>( this->args[1].getVariable()->getRevObject() ).getDagNode();
-    RevBayesCore::TypedDagNode<int>* numCats = static_cast<const Integer &>( this->args[2].getVariable()->getRevObject() ).getDagNode();
+    RevBayesCore::TypedDagNode<long>* numCats = static_cast<const Integer &>( this->args[2].getVariable()->getRevObject() ).getDagNode();
 
     RevBayesCore::DiscretizeLognormalQuadratureFunction *func = new RevBayesCore::DiscretizeLognormalQuadratureFunction( m, s, numCats);
